skin-analyzer/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
certifi
python-multipart